trv: Reordered gbmap traversal before mdir iteration
This is in preparation for some traversal simplification ideas, which
rely on all auxiliary/non-file btrees being visitable before file
btrees.

In theory the order of file vs auxiliary btrees doesn't really matter,
other than the number of different routes from mtree/mroot -> gbmap/file
btrees being a bit of a pain.

Note this is not true for the mtree, which must come first for
lfs3_mount to work.

---

Adds a bit of code when building with the gbmap:

                 code          stack          ctx
  before:       36480           2304          660
  after:        36476 (-0.0%)   2304 (+0.0%)  660 (+0.0%)

                 code          stack          ctx
  gbmap before: 39464           2320          772
  gbmap after:  39524 (+0.2%)   2320 (+0.0%)  772 (+0.0%)

                 code          stack          ctx
  gc before:    36552           2304          804
  gc after:     36548 (-0.0%)   2304 (+0.0%)  804 (+0.0%)