scripts: dbgflags.py: Renamed --o -> +o for prefix namespaces
Just a bit less typing than --o, and lowers risk of conflicts with
actual flags we may care about.

To be honest I was procrastinating because I thought this would be a lot
more work! I was prepared to write a hacky secondary parser, but argparse
already supports this natively with prefix_chars='-+'. Yay!