rattrs: Attempted to merge mtree split commits
The idea here was to merge mtree split commits to try to minimize
redundant logic that only differs in whether or not we need to create
the initial mtree weight.

Surprisingly, this backfired, adding more code than it saved. I guess I
underestimated how effective the compiler is at deduplicating these two
paths of logic:

                 code          stack          ctx
  before:       35316           2176          660
  after:        35324 (+0.0%)   2176 (+0.0%)  660 (+0.0%)

                 code          stack          ctx
  gbmap before: 38168           2192          772
  gbmap after:  38172 (+0.0%)   2192 (+0.0%)  772 (+0.0%)