preerase: Implemented the alloc-claim side of things (data blocks)
Note this is only half of preerased allocation.

And the easy half too.

The problem is littlefs's "restricted flash model" (as I'm now calling
it), which makes minimal assumptions about the behavior of the bd's
erase operation to support a wider range of devices. In particular,
littlefs doesn't assume the value of storage after an erase, which makes
detected failed progs (due to powerloss, etc) uniquely difficult.

For data blocks, we at least have the option of simply making sure the
relevant BMERASED range is deleted from the gbmap before use. This does
mean more progs during file writes, but in theory that is much cheaper
than erasing on-demand. And for storage where it's not, you should
probably consider not pre-erasing.

Fortunately, we don't actually need to commit to the gbmap to delete a
BMERASED range. If we consider BMERASED ranges outside the gbmap's known
window as invalid, we just need to decrement the known window to make
progress.

This is now implemented by lfs3_allocclaim, which forces an
lfs3_alloc_sync to ensure BMERASED blocks won't be reused even if power
is lost.

---

You may note this doesn't use the ecksums at all, except the check that
they're still valid. Unfortunately, we can't rely on ecksums for data
blocks because we don't control what gets progged. Worst-case, the data
being written matches the erased-state exactly, which is impossible for
littlefs to detect.

An alternative solution would be to add a header to every data block,
but this would come with several negatives:

- Headers would introduce some (albeit small) complexity into the write
  path, have limited value outside of preerases (we would still need to
  scan to allocate blocks), and raise questions around what headers
  should contain.

- Files would no longer perform optimally around powers-of-two, which
  may surprise users and risk unnecessarily poor performance.

- littlefs would lose its "universal migrator" status, as we would need
  to inject headers into any existing data blocks.

The in-gbmap ecksums solve the different problem of allocating metadata
blocks, which we can't use lfs3_allocclaim for as it would introduce
recursion.

---

Adds a large (but necessary) chunk of code/stack to the gbmap mode, but
only when preerasing (and some non-gbmap noise?):

                    code          stack          ctx
  before:          35116           2136          660
  after:           35112 (-0.0%)   2136 (+0.0%)  660 (+0.0%)

                    code          stack          ctx
  gbmap+np before: 38252           2144          776
  gbmap+np after:  38252 (+0.0%)   2144 (+0.0%)  776 (+0.0%)

                    code          stack          ctx
  gbmap+yp before: 38664           2144          796
  gbmap+yp after:  38832 (+0.4%)   2168 (+1.1%)  796 (+0.0%)