alloc: Moved ifdefs out of lfs3_alloc_can* functions
I'm not sure implicit ifdefs really help with readability.

They do redunce the number of lines, but the implicit ifdefs make it
harder to figure out what is actually compiled in.

The nice thing about explicit ifdefs is they're, well, explicit. This
makes it easier to rule out code paths early, and the earlier you can
rule things out, the easier it is to focus on what matters for the build
you care about.

---

TLDR IMO explicit ifdefs are preferable because that make it easier to
see what code is being compiled in to what build.

No code changes.