Adopted lfs3_wmask and friends
This just organizes the compat flags/masks a bit better, and avoids
needing to mess with the internals of lfs3_mountmroot anytime the wmask
flags change.

In case it isn't clear, the wmask/rmask/omask indicate which flags are
optional to mount the filesystem for the relevant mode. Currently the
only optional flag is LFS3_WCOMPAT_GBMAP.

Though, humorously, should lfs3_omask actually be all zeros?

---

Code changes minimal:

                    code          stack          ctx
  before:          35124           2136          660
  after:           35124 (+0.0%)   2136 (+0.0%)  660 (+0.0%)

                    code          stack          ctx
  gbmap+np before: 38264           2144          776
  gbmap+np after:  38252 (-0.0%)   2144 (+0.0%)  776 (+0.0%)

                    code          stack          ctx
  gbmap+yp before: 38844           2168          796
  gbmap+yp after:  38832 (-0.0%)   2168 (+0.0%)  796 (+0.0%)